// src/serializer/transform/model-types.ts
export interface ShapeModel {
  id: string;
  bpmnElement: string;
  bounds: BoundsModel;
  isExpanded?: boolean;
  isHorizontal?: boolean;
  label?: LabelModel;
}

export interface EdgeModel {
  id: string;
  bpmnElement: string;
  waypoints: PointModel[];
  label?: LabelModel;
}

export interface BoundsModel {
  x: number;
  y: number;
  width: number;
  height: number;
}

export interface PointModel {
  x: number;
  y: number;
}

export interface LabelModel {
  bounds?: BoundsModel;
}